Refactor Command type as enum
Removes unnecessary layer of abstraction by having a string Command type, making it easier to work with.
