+++ src/app/types.ts
@@ -1,6 +1,78 @@
-export type Command = string;
-
-export type Checksum = string;
+export enum Command {
+  POLL = "000",
+  STATUS_REPORT = "001",
+  NETWORK_LOGIN = "005",
+  COMMAND_ACKNOWLEDGE = "500",
+  COMMAND_ERROR = "501",
+  SYSTEM_ERROR = "502",
+  LOGIN = "505",
+  KEYPAD_LED_STATE = "510",
+  KEYPAD_LED_FLASH_STATE = "511",
+  ZONE_ALARM = "601",
+  ZONE_ALARM_RESTORE = "602",
+  ZONE_TAMPER = "603",
+  ZONE_TAMPER_RESTORE = "604",
+  ZONE_FAULT = "605",
+  ZONE_FAULT_RESTORE = "606",
+  ZONE_OPEN = "609",
+  ZONE_RESTORED = "610",
+  ENVISALINK_ZONE_TIMER_DUMP = "615",
+  BYPASSED_ZONES_BITFIELD_DUMP = "616",
+  DURESS_ALARM = "620",
+  F_KEY_ALARM = "621",
+  F_KEY_RESTORE = "622",
+  A_KEY_ALARM = "623",
+  A_KEY_RESTORE = "624",
+  P_KEY_ALARM = "625",
+  P_KEY_RESTORE = "626",
+  SMOKE_AUX_ALARM = "631",
+  SMOKE_AUX_RESTORE = "632",
+  PARTITION_READY = "650",
+  PARTITION_NOT_READY = "651",
+  PARTITION_ARMED = "652",
+  PARTITION_READY_FORCE_ARMING_ENABLED = "653",
+  PARTITION_IN_ALARM = "654",
+  PARTITION_DISARMED = "655",
+  EXIT_DELAY_IN_PROGRESS = "656",
+  ENTRY_DELAY_IN_PROGRESS = "657",
+  KEYPAD_LOCK_OUT = "658",
+  PARTITION_FAILED_TO_ARM = "659",
+  PGM_OUTPUT_IN_PROGRESS = "660",
+  CHIME_ENABLED = "663",
+  CHIME_DISABLED = "664",
+  INVALID_ACCESS_CODE = "670",
+  FUNCTION_NOT_AVAILABLE = "671",
+  FAILURE_TO_ARM = "672",
+  PARTITION_IS_BUSY = "673",
+  SYSTEM_ARMING_IN_PROGRESS = "674",
+  SYSTEM_IN_INSTALLERS_MODE = "680",
+  USER_CLOSING = "700",
+  SPECIAL_CLOSING = "701",
+  PARTIAL_CLOSING = "702",
+  USER_OPENING = "750",
+  SPECIAL_OPENING = "751",
+  PANEL_BATTERY_TROUBLE = "800",
+  PANEL_BATTERY_TROUBLE_RESTORE = "801",
+  PANEL_AC_TROUBLE = "802",
+  PANEL_AC_RESTORE = "803",
+  SYSTEM_BELL_TROUBLE = "806",
+  SYSTEM_BELL_TROUBLE_RESTORE = "807",
+  FTC_TROUBLE = "814",
+  FTC_TROUBLE_RESTORE = "815",
+  BUFFER_NEAR_FULL = "816",
+  GENERAL_SYSTEM_TAMPER = "829",
+  GENERAL_SYSTEM_TAMPER_RESTORE = "830",
+  TROUBLE_LED_ON = "840",
+  TROUBLE_LED_OFF = "841",
+  FIRE_TROUBLE_ALARM = "842",
+  FIRE_TROUBLE_ALARM_RESTORE = "843",
+  VERBOSE_TROUBLE_STATUS = "849",
+  CODE_REQUIRED = "900",
+  COMMAND_OUTPUT_PRESSED = "912",
+  MASTER_CODE_REQUIRED = "921",
+  INSTALLERS_CODE_REQUIRED = "922",
+  SOFTWARE_ZONE_ALARM = "S01",
+}
 
 export type Data = {
   partition: number;
@@ -17,5 +89,5 @@ export type Event = {
 export type Payload = {
   command: Command;
   data: Data;
-  checksum: Checksum;
+  checksum: string;
 };
